fix: rename tera-model-configuration-ciemss to tera-model-configuration-distribution
