+++ packages/client/hmi-client/src/components/models/tera-model-configuration-distribution.vue
